Further *BSD portability.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@5958 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/DebuggerWX/Src/MemoryWindow.cpp
@@ -324,7 +324,7 @@ void CMemoryWindow::onSearch(wxCommandEvent& event) {
 	std::vector<u8> Dest;//May need a better name
 	u32 size=0;
 	int pad=rawData.size()%2;//If it's uneven
-	unsigned long i=0;
+	unsigned int i=0;
 	long count=0;
 	char copy[3]={0};
 	long newsize=0;
